Extract queries generation (before reuse in another script)
